BAEL-3287: rename the package to apply review note
